+++ crankshaft-engine/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+* `Execution`s can now specify multiple container images to act as fallbacks ([#83](https://github.com/stjude-rust-labs/crankshaft/pull/83)).
+* `Backend::run()` now returns a collection of `ExecutionResult`s, rather than `ExitStatus` ([#83](https://github.com/stjude-rust-labs/crankshaft/pull/83)).
+
 ## 0.10.0 - 04-22-2026
 
 ### Added